MDL-82654 quiz: Retain backwards compatibility in web services
To ensure backwards compatibility, particularly with the mobile app, the
mod_quiz_get_attempt_review function will continue to return only the
'finished' and 'abandoned' states for quiz attempts.

'submitted' attempts will be reported as 'finished'.

This function is now deprecated. Is it replaced by the new route
`/mod_quiz/attempts/{id}/review`. This is defined in
`mod_quiz\route\api\attempts\get_review`, which follows the original
logic of the external function so will return the true attempt state.