MDL-83862 question: Improve error handling of question bank filters.
This provides better handing of errors from question bank filters
by displaying messages from thrown exceptions in an alert,
allowing the error messages to be tested for. This allows filter
conditions to perform some validation and return an error if
something is not valid.

It also moves the code updating the URL parameters to after the
filtered questions have been successfully returned, meaning that
if there is an error, the invalid values will never be set in the
URL. This avoids users accidentally sharing a URL that will always
display an error.