MDL-85517 course: Reset large courses asynchronously
Very large courses can take a long time to reset due to large number of
users to unenrol, large numbers of grades to delete, etc.

This moves the processing of a course reset to an ad-hoc task for
courses over a certain threshold, currently based on the number of
enrolments. A task indicator is displayed on the Reset page to show
progress of the task, and the user is redirected back to the course page
once the task is complete.