MDL-85069 question: Services for question bank and category search
This updates the core_question_search_shared_banks external function
with additional parameters to make it applicable outside of the "Switch
bank" dialogue in the quiz, with the corresponding changes to the
question_banks_datasource Javascript module, and unit test coverage.

It also adds a new core_question_search_question_categories external
function, for searching question categories within a particular bank,
and a question_categories_datasource Javascript module.

These are both for use with autocomplete elements.