MDL-85975 backup: Exclude null values in questiondata
unset_excluded_fields() was using isset() to determine if the field
targeted for removal is present in the provided data structure. However,
isset() returns false if the field exists, but contains null. This means
the field will not be unset when it should be.

This resolves this by changing the isset() to a property_exists() check.
It also corrects the type hint on the $data parameter, as it should
always be an object - the method does not handle arrays.