MDL-87103 questions: Replace "display" methods
The question bank was being output with a stack of functions which each echoed
fragments of HTML, much of it generated with html_writer. At the top level,
these were wrapped in output buffers to return the HTML rather than outputting
it directly. However, situations could occur where one of these functions threw
an exception after output had started, resulting in messy output or invalid
JSON from AJAX requests.

This completely replaces the existing functions with templatables and
templates, so that all functions producing output for the question bank
now return HTML rather than rendering it. The question bank view class
itself is now templatable, and all the templates for the constituent
components of the view (filters, paging bars, question tables, question
rows, bulk actions) are included as partial templates.

At the interface with qbank plugins, it is impossible to nest the plugin templates
inside the core templates, so instead we expect plugins return rendered HTML
output for things like table cell content, and include that HTML verbatim in the
core templates.

This deprecates a lot of functions from the core_question\local\bank\view class.