MDL-87103 qbank_customfields: Update custom field cell
This replaces the `display_content` method of the
`custom_field_column` class with a `render`
method, to return the HTML for a custom field cell
using a template, rather than echoing it directly.

This also deprecates the `table_display` method in the
plugin's renderer, in favour of including this template
as a partial in the cell's template.