MDL-87103 qbank_editquestion: Update question status cell
This replaces the `display_content` method of the
`question_status_column` class with a `render`
method, to return the HTML for a question status cell
using a template, rather than echoing it directly.

This also deprecates the `render_status_dropdown` method
in the plugin's renderer, in favour of including this
template as a partial in the cell's template.

It also optimises the Javascript for displaying the modal
by attaching a single event listener to the table, rather
than attaching one to each cell.