MDL-87103 qbank_history: Update cell and view
This replaces the `display_content` method of the
`version_number_column` class with a `render`
method, to return the HTML for a version number cell
using a template, rather than echoing it directly.

It also updates the `question_history_view` class,
a customised question bank view for displaying all
versions of a single question. This view displays
a custom header and a question list, but no filters.
It is now using the templatable interface inherited
from the parent view.