MDL-87103 qbank_usage: Update cells
This replaces the `display_content` method of the
`question_usage_column` and `question_last_used_column`
classes with a `render` method, to return the HTML
for a these cells using a template, rather than
echoing it directly.

It also deprecates the `render_last_used_column` method
in the plugin's renderer, as this is now handled by
a template.