MDL-87103 qbank_viewcreator: Update creator and modifier cells
This replaces the `display_content` method of the
`create_name_column` and `modifier_name_column`
classes with a `render` method, to return the HTML
for a these cells using a template, rather than
echoing it directly.

These columns both use the new `name_cell`
templatable and template for output, as they
both display a name and date in the same format.
This deprecates the existing `creator_display`
template.